+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 # indexing metrics collector Changelog
 
 ## [Unreleased]
+### Changed
 
-## [1.0.0]
 ### Added
 - Initial Release under elastic
 - Rebranded from initial spike 
@@ -12,7 +12,10 @@
   - [Access token authentication](https://github.com/breskeby/indexing-stats-collector/issues/3)
   - [API keys authentication](https://github.com/breskeby/indexing-stats-collector/issues/2)
 - [Make elasticsearch index configurable](https://github.com/breskeby/indexing-stats-collector/issues/5)
+- add runtime data about 
 
 ### Changed
-- Do not use JDK 8 specific class javax.xml.bind.DatatypeConverter in user data anymizer
+- Renamed default index name to `idea-metrics-indexing`
+- Improve elasticsearch connection error handling
+- Do not use JDK 8 specific classes in user data anymizer